+++ sound-mind-studio/tests/test_canvas_widget.h
@@ -26,4 +26,6 @@ private slots:
     void showBoundingBoxesDrawsNothingWhenOff();
     void showBoundingBoxesDrawsAnActiveOperationsBoundingBox();
     void showPathGeometryDrawsAnActiveOperationsPath();
+    void mouseMoveEmitsCursorMovedRegardlessOfToolMode();
+    void leavingTheCanvasEmitsCursorLeft();
 };